refactor: move metadata to separate folder
